fix: added needed QR.cu for cuda implementation of PhD dissertation code chore: deleted unneeded .DS_Store files created by mac file manager
The file QR.cu is the source implementation of the QR factorization
algorithm in Cuda for use in Benettin's algorithm (computing Lyapunov
exponents).  I had forgotten to include it in the original implementation,
so I'm including it now.  The cleanup of all of the .DS_Store files was
done to make the repo cleaner and easier to navigate.

Resolves: N/A
See also: N/A